refactor, variable update
